# builds/firebase-auth/dashboard/src/app/api/batch-verify/route.ts
import { callService } from '@/lib/api';
import { NextRequest, NextResponse } from 'next/server';

export async function POST(request: NextRequest) {
  let body: unknown;
  try {
    body = await request.json();
  } catch {
    return NextResponse.json({ error: 'Invalid JSON' }, { status: 400 });
  }

  try {
    const res = await callService('/batch-verify', { method: 'POST', body });
    const data = await res.json();
    return NextResponse.json(data, { status: res.status });
  } catch {
    return NextResponse.json(
      { error: 'Service unreachable' },
      { status: 502 }
    );
  }
}